+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## UNRELEASED
+
+- Add `--exclude-response-headers` CLI option.  This option will suppress response
+headers from the dumper output.
+
 ## 5.3.0 (2023-03-07)
 
 - Add `--verbose` CLI option.  This option will print the dumper and block names.